reorg files, move to icon to right with hover text
